#170
* implement QminderAPIProtocol
* protocol extension for QminderAPIProtocol
* convert QminderAPI to struct and conform to QminderAPIProtocol
* use QminderAPIProtocol in tests
